requirements.txt
numpy
torch
numpy-stl
torchvision
scikit-learn